effects: share sandbox's and catch's handlers, not just their names
Declaring an operation in a shared layer while each host writes its own
copy of the obvious implementation would leave the JavaScript exactly
where it was. `sandbox`'s clock helper needs a `try` and a
`performance.now()`, and neither is any particular host's — a browser has
both — so it moves to `common/module.mjs` beside the declarations, with
the two handlers as an object a host spreads into its own map.

`effects/node`'s runner spreads it and its local copy is gone. A browser
interpreter spreads the same object, which is the point of it being one.

`effects/browser` is deliberately not created yet: today it would be one
line wiring handlers nobody calls, since the page does not run through
effects until the traversal port. It lands with its first consumer —
the same second-implementer discipline that decided what moved here.
